requirements.txt
# step01
psycopg2-binary
python-dotenv
sqlalchemy
# step02
pymysql
cryptography



